Remove resolver dependency on core
