+++ crates/parcel_core/Cargo.toml
@@ -9,9 +9,9 @@ default = []
 
 [dependencies]
 parcel_filesystem = { path = "../parcel_filesystem" }
-parcel-resolver = { path = "../../packages/utils/node-resolver-rs" }
 
 anyhow = "1.0.82"
+bitflags = "1.3.2"
 browserslist-rs = "0.15.0"
 dyn-hash = "0.x"
 nodejs-semver = "4.0.0"
